feat: ft-wave1-gemini-golden-004 - Consume migration-ledger rows and close verification gates
Move the nine Gemini process-harness scenarios into the workers/inference/gemini
golden cell, update migration-ledger source paths, and mark the checklist complete.
