+++ docs/temp/functional-tests-expansion/test-file-checklist.md
@@ -290,10 +290,19 @@ under `work/`, `sessions/`, `factory/`, and `product/`.
   - `TestOpenCodeGoldenFinalOnlyFallback`.
   - `TestOpenCodeGoldenStructuredFailureAndTimeout`.
 
-- [ ] `tests/functional/workers/inference/gemini/golden_test.go`
+- [x] `tests/functional/workers/inference/gemini/golden_test.go`
   - `TestGeminiGoldenTextSuccess`.
   - `TestGeminiGoldenRateLimitAndStructuredFailure`.
   - `TestGeminiGoldenTimeout`.
+  - `TestRootBuiltProcessExecutesThroughSharedSupport`.
+  - `TestGeminiConductorSuccessThroughRootBuildProcess`.
+  - `TestGeminiClassifierRejectsStructuredLabelThroughRootBuildProcess`.
+  - `TestGeminiConductorPreservesConfiguredEnvironment`.
+  - `TestGeminiConductorPreservesConfiguredSkipPermissions`.
+  - `TestGeminiRejectsUnsupportedStructuredOutputBeforeProviderIO`.
+  - `TestGeminiRejectsUnsupportedWorkingDirectoryBeforeProviderIO`.
+  - `TestGeminiNativeFailureThroughRootBuildProcessIsSafe`.
+  - `TestGeminiCommandCancellationThroughRootBuildProcessIsCanonical`.
 
 - [ ] `tests/functional/workers/inference/kiro/golden_test.go`
   - `TestKiroGoldenTextSuccess`.
